fix: address PR #849 review pass 2 (instrumentation claim was backwards)
- BLOCKING: the streamed-span materiality bound claimed instrumentation
  re-wraps the model and routes to the eager path. Backwards on 2.31.0:
  instrumentation is capability-based and explicit InstrumentedModels are
  unwrapped before the run, so instrumented requests carry the bare
  round-robin model and take the OWNED path - reachable under the shipped
  logfire.instrument_pydantic_ai() (observability.py). Docstring rewritten
  honestly (scope of the loss: only the leaf-attribute refinement on a
  torn-down stream's chat span) and pinned by a new test proving the
  identity gate HOLDS under real InstrumentationSettings.
- NON-BLOCKING: 'every request advances exactly once' custody wording
  disambiguated (wrapped request vs continuation segment) and the guest
  list corrected (arbitrary WrapperModel, explicitly NOT instrumentation).
- Split the test file (750 lines > 600 cap) into seam-contract +
  runs/wiring files sharing round_robin_capability_harness.py.
